Update func utils and tests to not use regex
