+++ src/index.ts
@@ -15,8 +15,6 @@ import { AzureLoginPlugin } from "./plugins/login/loginPlugin";
 import { AzureApimServicePlugin } from "./plugins/apim/apimServicePlugin";
 import { AzureApimFunctionPlugin } from "./plugins/apim/apimFunctionPlugin";
 import { AzureFuncPlugin } from "./plugins/func/azureFunc";
-import { AzureFuncAddPlugin } from "./plugins/func/add/azureFuncAdd";
-import { AzureFuncRemovePlugin } from "./plugins/func/remove/azureFuncRemove";
 
 
 export class AzureIndex {
@@ -34,8 +32,6 @@ export class AzureIndex {
     this.serverless.pluginManager.addPlugin(AzureApimServicePlugin);
     this.serverless.pluginManager.addPlugin(AzureApimFunctionPlugin);
     this.serverless.pluginManager.addPlugin(AzureFuncPlugin);
-    this.serverless.pluginManager.addPlugin(AzureFuncAddPlugin);
-    this.serverless.pluginManager.addPlugin(AzureFuncRemovePlugin);
   }
 }
